gdxInfo {gdxrrw}	R Documentation

Display or Return Information from a GDX File

Description

Display all the information contained in a GDX file, along with some version & debugging info. The display format is copied from gdxdump.

Alternately, return information about the symbols contained in a GDX file in list or a more extensive data frame format.

Usage

  gdxInfo(gdxName = NULL, dump=TRUE, returnList=FALSE, returnDF=FALSE)

Arguments

gdxName
the name of the GDX file. If this argument is omitted, we display only the version info for the GDX library used

dump
if TRUE, dump GDX contents to the console

returnList
if TRUE, return brief information about the symbols in the GDX file as a list of lists

returnDF
if TRUE, return extended information about the symbols in the GDX file as a list of data frames

Note

A common problem is failure to load the external GDX libraries that are required to interface with GDX data. Use igdx to troubleshoot and solve this problem.

Author(s)

Original coding by Rishabh Jain. Adopted and packaged by Steve Dirkse. Maintainer: R@gams.com

See Also

igdx, rgdx, wgdx

Examples

  gdxInfo();
  ## Not run: 
    gdxInfo("trnsport.gdx");
    gdxInfo("trnsport.gdx", dump=FALSE, returnDF=TRUE);
  
## End(Not run)

[Package gdxrrw version 1.0.6 Index]