gdxrrw-package {gdxrrw}	R Documentation

Exchanging data between GAMS and R

Description

This package implements an interface between GAMS and R. It includes functions to transfer data between GDX (the GAMS data format) and R, a function to call GAMS from within R, query functions for the meta-data in GDX, and other related utilities.

Details

This package contains five classes of functions:

igdx(...) gives information on the linkage between this package and the GDX library.

rgdx("gdxfile", ...) and related functions read from GDX

wgdx("gdxfile", ...) and related functions write to GDX

gams("gmsfile and args", ...) runs gams with the arguments provided

gdxInfo("gdxfile", ...) dumps GDX content or returns GDX metadata (list of symbols, etc.)

Author(s)

Original coding by Rishabh Jain. Adopted, packaged, and extended by Steve Dirkse. Maintainer: R@gams.com

[Package gdxrrw version 1.0.6 Index]