wgdx {gdxrrw}	R Documentation

Write R data to GDX

Description

Create a GDX file containing the GAMS data described in the lists .... Note that each list describes a separate data element - i.e. a set or parameter - all of which are written to a single file.

Usage

  # generic form - each arg is a list specifying a symbol to write
  wgdx(gdxName, ..., squeeze = 'y')

  # write multiple symbols specified in list, data frame, or scalar form
  wgdx.lst(gdxName, ..., squeeze = 'y')

Arguments

gdxName
the name of the GDX file to write

...
zero or more arguments describing the symbols to write. For wgdx, each arg must be a list holding a symbol (e.g. as returned by rgdx). For wgdx.lst, each arg is also allowed to be a data frame holding a symbol (e.g. as returned by rgdx.param or rgdx.set), a scalar holding a symbol (e.g. as returned by rgdx.scalar), or a list containing any combination of these elements.

squeeze
if 'y'/TRUE/nonzero, squeeze out zeros: do not store in GDX. If 'n'/FALSE/zero, do not squeeze out zeros: store explicit zeros in GDX. If 'e', store zeros as EPS in the GDX.

Note

A common problem is failure to load the external GDX libraries that are required to interface with GDX data. Use igdx to troubleshoot and solve this problem.

Author(s)

Original coding by Rishabh Jain. Adopted and packaged by Steve Dirkse. Maintainer: R@gams.com

See Also

igdx, rgdx, gdxInfo

Examples

  data(trnsport);
  wgdx("wgdx1.gdx",sf,si,sj,sa,sb,sd);

  ## Not run: 
    # complete tests and examples can be run in the
    # extdata directory of the gdxrrw package
    # check .libPaths for a hint on where packages are installed
    setwd(paste(.libPaths()[1],"/gdxrrw/extdata",sep=""))
    source("tAll.R")
  
## End(Not run)

[Package gdxrrw version 1.0.6 Index]